+++ src/main/java/com/example/csemaster/dto/QuizUpdateDTO.java
@@ -1,4 +1,4 @@
-package com.example.csemaster.features.quiz.minyoung;
+package com.example.csemaster.dto;
 
 import lombok.Data;
