+++ src/main/java/com/example/csemaster/features/quiz/QuizCreateController.java
@@ -3,8 +3,6 @@ package com.example.csemaster.features.quiz;
 import com.example.csemaster.dto.QuizDTO;
 import com.example.csemaster.dto.request.QuizImageRequest;
 import com.example.csemaster.dto.request.QuizReportRequest;
-import com.example.csemaster.dto.request.QuizSolverRequest;
-import com.example.csemaster.dto.response.QuizResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.Authentication;
@@ -15,28 +13,10 @@ import org.springframework.web.bind.annotation.*;
 @RestController
 @RequiredArgsConstructor
 @RequestMapping("/api/quiz")
-public class QuizController {
+public class QuizCreateController {
     private final QuizSolverService quizSolverService;
     private final QuizCreateService quizCreateService;
 
-
-    @GetMapping("/random")
-    public QuizResponse getRandomQuiz(@RequestParam String subject, @RequestParam String detailSubject) {
-        // 사용자 인증 정보 가져오기
-        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-        String userId = authentication.getName();
-
-        // 무작위로 하나의 문제를 반환
-        return quizSolverService.getQuiz(userId, subject, detailSubject);
-    }
-
-    @PostMapping("/solver")
-    public ResponseEntity<?> solveQuiz(@RequestBody QuizSolverRequest request)  {
-        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-        String userId = authentication.getName();
-
-        return quizSolverService.saveQuizResult(userId, request.getQuizId(), request.getIsCorrect());
-    }
 
     @PostMapping("/report")
     public ResponseEntity<?> quizReport(@RequestBody QuizReportRequest request) {
