+++ src/main/java/com/example/csemaster/features/quiz/QuizManagerService.java
@@ -1,20 +1,40 @@
-package com.example.csemaster.features.quiz.minyoung;
+package com.example.csemaster.features.quiz;
 
+import com.example.csemaster.dto.QuizUpdateDTO;
 import com.example.csemaster.entity.QuizEntity;
-import com.example.csemaster.features.quiz.QuizCreateService;
 import com.example.csemaster.repository.QuizRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.Optional;
 
 @Service
 @Slf4j
 @RequiredArgsConstructor
-public class QuizUpdateService {
+public class QuizManagerService {
+
     private final QuizRepository quizRepository;
     private final QuizCreateService quizCreateService;
+
+    @Transactional
+    public boolean deleteQuiz(Long quizId) {
+        Optional<QuizEntity> quiz = quizRepository.findByQuizId(quizId);
+
+        // 존재하는 quiz인지 확인
+        if (!quiz.isPresent()) {
+            throw new RuntimeException("Incorrect quizId");
+        }
+
+        quiz.get().setIsDeleted(true);
+        quizRepository.save(quiz.get());
+
+        log.info("Delete Successfully");
+
+        return true;
+    }
+
     public boolean updateQuiz(QuizUpdateDTO quizUpdateDTO) {
         Optional<QuizEntity> quiz = quizRepository.findByQuizId(quizUpdateDTO.getQuizId());
