+++ src/main/java/com/example/csemaster/features/quiz/QuizReportController.java
@@ -1,5 +1,6 @@
-package com.example.csemaster.features.quiz.minyoung;
+package com.example.csemaster.features.quiz;
 
+import com.example.csemaster.dto.response.QuizReportResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
